@@ -6,22 +6,8 @@ import java.util.Collection;
  * @author 蒋时华
  * @date 2020-05-02 16:42:27
  */
-public interface IDataPageCollectionResult<T> extends IDataPageResult<Collection<T>> {
+public interface IDataPageCollectionResult<T> extends IPageResult<Collection<T>> {
 
-    /**
-     * get page list data
-     *
-     * @return
-     */
-    @Override
-    Collection<T> getData();
 
-    /**
-     * set page list data
-     *
-     * @param list
-     */
-    @Override
-    void setData(Collection<T> list);
 
 }
