@@ -8,22 +8,7 @@ import java.util.HashSet;
  * @author 蒋时华
  * @date 2018/7/19
  */
-public interface IDataPageHashSetResult<T> extends IDataPageResult<HashSet<T>> {
+public interface IDataPageHashSetResult<T> extends IPageResult<HashSet<T>> {
 
-    /**
-     * get response data
-     *
-     * @return T
-     */
-    @Override
-    HashSet<T> getData();
-
-    /**
-     * set response data
-     *
-     * @param data
-     */
-    @Override
-    void setData(HashSet<T> data);
 
 }
