@@ -1,13 +1,16 @@
 package com.skrstop.framework.components.example.starters.simple.controller;
 
 import com.skrstop.framework.components.core.common.response.MapResult;
-import com.skrstop.framework.components.core.common.response.PageLinkedSetResult;
+import com.skrstop.framework.components.core.common.response.PageListResult;
 import com.skrstop.framework.components.core.common.response.core.IPageResult;
 import com.skrstop.framework.components.core.common.response.core.IResult;
+import com.skrstop.framework.components.core.common.response.page.ListSimplePageData;
+import com.skrstop.framework.components.core.common.response.page.SimplePageData;
 import com.skrstop.framework.components.example.starters.simple.service.Example1MysqlService;
 import com.skrstop.framework.components.starter.id.service.IdService;
 import com.skrstop.framework.components.starter.web.response.DisableGlobalResponse;
 import com.skrstop.framework.components.starter.web.response.DisableTransResultTypeResponse;
+import com.skrstop.framework.components.util.value.data.CollectionUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
@@ -85,7 +88,22 @@ public class ExampleReturnController {
      */
     @GetMapping("/exampleReturnPage")
     public IPageResult exampleReturnPage() {
-        return PageLinkedSetResult.Builder.success(10, 100);
+        return PageListResult.Builder.success(10, 100, 0, CollectionUtil.newArrayList("a", "b"));
+    }
+
+    /**
+     * page样例
+     *
+     * @return
+     */
+    @GetMapping("/exampleReturnPage2")
+    public SimplePageData exampleReturnPage2() {
+        ListSimplePageData<String> result = new ListSimplePageData<>();
+        result.setPageNumber(10);
+        result.setPageSize(100);
+        result.setTotal(0);
+        result.setRows(CollectionUtil.newArrayList("a", "b"));
+        return result;
     }
 
 }
