@@ -22,6 +22,9 @@ import java.util.HashSet;
 @Order(Ordered.HIGHEST_PRECEDENCE)
 public class GlobalFeignProperties {
 
+    /*** 是否开启http2 */
+    private boolean enableHttp2 = false;
+
     /**
      * feign扫描包路径，多个路径逗号分隔，也可以使用@EnableFeignClients
      */
@@ -32,11 +35,11 @@ public class GlobalFeignProperties {
      * 如果不使用Result包装, 则只返回response中的data部分，访问失败则返回null
      * 如果使用Result包装, 则只返回有Result包装的对象
      * <p>
-     * 注意：如果开启了自适应处理，分页接口则返回 CommonPageData
+     * 注意：如果开启了自适应处理，分页接口则返回 SimplePageData
      * 注意：此参数为客户端配置，服务端配置可在starter-web中关闭 useFeignSupport 可实现服务端不返回包装格式
      * 当服务端关闭userFeignSupport时，客户端需要开启此参数，否则将类型转换错误，导致报错
      *
-     * @see com.skrstop.framework.components.core.common.response.page.CommonPageData
+     * @see com.skrstop.framework.components.core.common.response.page.SimplePageData
      */
     private boolean stopAutoRemoveGlobalResponse = false;
 
