@@ -307,7 +307,7 @@ public class DynamicRedissonClient {
      * @param codec codec for values
      * @return JsonBucket object
      */
-    public <V> RJsonBucket<V> getJsonBucket(String dsKey, String name, JsonCodec<V> codec) {
+    public <V> RJsonBucket<V> getJsonBucket(String dsKey, String name, JsonCodec codec) {
         RedissonClient redissonClient = this.multiRedisson.getRedissonClient();
         return redissonClient.getJsonBucket(name, codec);
     }
