@@ -23,6 +23,7 @@ import lombok.SneakyThrows;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.InvalidMediaTypeException;
 import org.springframework.http.MediaType;
+import org.springframework.util.MimeTypeUtils;
 import org.springframework.web.reactive.function.server.ServerResponse;
 import org.springframework.web.server.ServerWebExchange;
 import reactor.core.publisher.Mono;
@@ -30,7 +31,7 @@ import reactor.core.publisher.Mono;
 import java.io.IOException;
 import java.util.List;
 
-import static org.springframework.web.reactive.function.BodyInserters.fromObject;
+import static org.springframework.web.reactive.function.BodyInserters.fromValue;
 
 /**
  * The default implementation of {@link BlockRequestHandler}.
@@ -50,14 +51,14 @@ public class DefaultWebFluxBlockRequestHandler implements BlockRequestHandler {
 //        }
         // JSON result by default.
         return ServerResponse.status(HttpStatus.TOO_MANY_REQUESTS)
-                .contentType(MediaType.APPLICATION_JSON_UTF8)
-                .body(fromObject(buildErrorResult(ex, exchange)));
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(fromValue(buildErrorResult(ex, exchange)));
     }
 
     private Mono<ServerResponse> htmlErrorResponse(Throwable ex) {
         return ServerResponse.status(HttpStatus.TOO_MANY_REQUESTS)
                 .contentType(MediaType.TEXT_PLAIN)
-                .syncBody(DEFAULT_BLOCK_MSG_PREFIX + ex.getClass().getSimpleName());
+                .bodyValue(DEFAULT_BLOCK_MSG_PREFIX + ex.getClass().getSimpleName());
     }
 
     private String buildErrorResult(Throwable ex, ServerWebExchange exchange) throws IOException {
@@ -71,7 +72,7 @@ public class DefaultWebFluxBlockRequestHandler implements BlockRequestHandler {
         try {
             List<MediaType> acceptedMediaTypes = exchange.getRequest().getHeaders().getAccept();
             acceptedMediaTypes.remove(MediaType.ALL);
-            MediaType.sortBySpecificityAndQuality(acceptedMediaTypes);
+            MimeTypeUtils.sortBySpecificity(acceptedMediaTypes);
             return acceptedMediaTypes.stream()
                     .anyMatch(MediaType.TEXT_HTML::isCompatibleWith);
         } catch (InvalidMediaTypeException ex) {
